+++ CHANGELOG.md
@@ -2,6 +2,9 @@
 
 # 2.0.7
 - Add support for SuspendReporting in Android Platform
+- Upgrade iOSAgent to 1.8.7, android-agent to 6.1.0
+- Android: Added queryTrackedDomainList feature to exclude parameters from url other than in this list
+- iOS: Added queryTrackedDomainList feature to exclude parameters from url other than in this list
 
 # 2.0.6
 - Support Typescript projects